Trenne Funktionsdefinitionen in cpp-Dateien
