+++ src/wifi_manager.cpp
@@ -1,0 +1,106 @@
+#include "wifi_manager.h"
+
+// Funktionen aus wifi_manager.h implementiert
+
+void clearWiFiSymbol() {
+    if (!triggerActive) {
+        Serial.println(F("🚫 WiFi-Symbol wird entfernt."));
+        display.fillScreen(display.color565(0, 0, 0));
+        display.display();
+    } else {
+        Serial.println(F("⏳ WiFi-Symbol bleibt, weil ein Buchstabe aktiv ist."));
+    }
+}
+
+#define SCALE_FACTOR 2
+
+void drawWiFiSymbol() {
+    if (wifiConnected) {
+        Serial.println(F("📶 WiFi-Symbol wird angezeigt."));
+
+        if (!triggerActive) {
+            display.fillScreen(display.color565(0, 0, 255));
+
+            int x_pos = (64 - (32 * SCALE_FACTOR)) / 2;
+            int y_pos = (64 - (32 * SCALE_FACTOR)) / 2;
+
+            for (int y = 0; y < 32; y++) {
+                for (int x = 0; x < 32; x++) {
+                    uint8_t rowValue = pgm_read_byte(&letterData['~'][y * 4 + (x / 8)]);
+                    if (rowValue & (1 << (7 - (x % 8)))) {
+                        display.fillRect(x_pos + x * SCALE_FACTOR, y_pos + y * SCALE_FACTOR, SCALE_FACTOR, SCALE_FACTOR, display.color565(0, 0, 0));
+                    }
+                }
+            }
+            display.display();
+        } else {
+            Serial.println(F("⏳ WiFi-Symbol NICHT angezeigt, weil ein Buchstabe aktiv ist."));
+        }
+    }
+}
+
+void connectWiFi() {
+    Serial.println(F("🌐 Verbinde mit WiFi..."));
+    WiFi.hostname(hostname);
+    WiFi.begin(wifi_ssid, wifi_password);
+
+    unsigned long startAttempt = millis();
+    unsigned long lastDot = startAttempt;
+    while (WiFi.status() != WL_CONNECTED &&
+           (millis() - startAttempt < (wifi_connect_timeout * 1000UL))) {
+        if (millis() - lastDot >= 2000) {
+            Serial.print(".");
+            lastDot = millis();
+        }
+        delay(10);
+    }
+
+    if (WiFi.status() == WL_CONNECTED) {
+        Serial.println(F("\n✅ WiFi verbunden!"));
+        Serial.print(F("IP-Adresse: "));
+        Serial.println(WiFi.localIP());
+        wifiConnected = true;
+        drawWiFiSymbol();
+        setupWebServer();
+    } else {
+        Serial.println(F("\n⛔ WiFi Timeout! Verbindung fehlgeschlagen. WiFi bleibt aus."));
+        wifiConnected = false;
+        WiFi.disconnect();
+        WiFi.mode(WIFI_OFF);
+    }
+}
+
+void checkWiFi() {
+    if (WiFi.status() != WL_CONNECTED) {
+        if (!wifiDisabled) {
+            Serial.println(F("⚠️ WLAN-Verbindung verloren. Schalte WiFi & Webserver aus..."));
+            wifiConnected = false;
+
+            if (!triggerActive) {
+                clearWiFiSymbol();
+            }
+
+            WiFi.disconnect();
+            WiFi.mode(WIFI_OFF);
+
+            server.end();
+            Serial.println(F("🌐 Webserver gestoppt. Neustart erforderlich für neue Verbindung."));
+
+            wifiDisabled = true;
+        }
+    } else {
+        if (!wifiConnected) {
+            Serial.println(F("✅ WLAN verbunden!"));
+            Serial.print(F("IP-Adresse: "));
+            Serial.println(WiFi.localIP());
+
+            wifiConnected = true;
+            wifiDisabled = false;
+
+            if (!triggerActive) {
+                drawWiFiSymbol();
+            }
+        }
+    }
+}
+
